+++ zimui/src/components/channel/tabs/LivesTab.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import TabView from '@/views/TabView.vue'
+import GenericTab from './GenericTab.vue'
 </script>
 
 <template>
-    <tab-view :playlist-slug="'userLivesPlaylist'" title-prefix="Lives" />
+    <generic-tab :playlist-slug="'userLivesPlaylist'" title-prefix="Lives" />
 </template>