+++ zimui/src/components/channel/tabs/ShortsTab.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import TabView from '@/views/TabView.vue'
+import GenericTab from './GenericTab.vue'
 </script>
 
 <template>
-  <tab-view :playlist-slug="'userShortUploadsPlaylist'" title-prefix="Shorts" />
+  <generic-tab :playlist-slug="'userShortUploadsPlaylist'" title-prefix="Shorts" />
 </template>